requirements.txt
asgiref==3.6.0
Django==4.1.7
django-autoslug==1.9.8
django-tinymce==3.6.1
Pillow==9.4.0
python-slugify==8.0.1
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2022.7
